+++ CHANGELOG.md
@@ -87,3 +87,5 @@
 * Define `U_1_II` data type
 * Define `U_I_UU_III_U_II_I` data type
 * Define `Reference` type alias
+* Define `W_I_II_II` data type
+* Define `Attribute` type alias